Add serialized pipelines with legacy External Source
The pipelines were created with legacy External Source
implementation using DALI v1.0.0.

The following code was used:
{
  Pipeline p(10, 4, 0);
  p.AddExternalInput("es");
  p.Build({{"es", "cpu"}});
  std::ofstream file("add_external_input_v1.0.0.dali",
                      std::ios::out | std::ios::binary);
  file << p.SerializeToProtobuf();
}
for (auto dev : {"cpu"s, "gpu"s}) {
  Pipeline p(10, 4, 0);
  p.AddOperator(OpSpec("_ExternalSource")
                    .AddArg("device", dev)
                    .AddArg("name", "es")
                    .AddOutput("es", dev),
                "es");
  p.Build({{"es", dev}});
  std::ofstream file("underscore_ext_src_" + dev + "_v1.0.0.dali",
                      std::ios::out | std::ios::binary);
  file << p.SerializeToProtobuf();
}
With DALI up to 1.3 those Pipelines would serialize
differently than how later versions of DALI do it,
but we want to ensure that they could still be loaded
and the result would be the same.

Signed-off-by: Krzysztof Lecki <klecki@nvidia.com>